.claude/worktrees/sad-haslett/hackathons/h4-cloud-native/services/discord-bot/requirements.txt
discord.py==2.3.2
aiohttp==3.9.3
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.23.3
aioresponses==0.7.6
